# Speed up Array.Reverse by using ref reassignment