--- a/src/mscorlib/src/System/Array.cs
+++ b/src/mscorlib/src/System/Array.cs
@@ -1559,7 +1559,11 @@ namespace System
         {
             if (array == null)
                 ThrowHelper.ThrowArgumentNullException(ExceptionArgument.array);
-            Reverse(array, 0, array.Length);
+
+            // The Span ctor might perform an unnecessary type check if T is a reference type,
+            // so we use MemoryMarshal.CreateSpan to suppress this check.
+
+            MemoryMarshal.CreateSpan(ref Unsafe.As<byte, T>(ref array.GetRawSzArrayData()), array.Length).Reverse();
         }
 
         public static void Reverse<T>(T[] array, int index, int length)
@@ -1571,19 +1575,7 @@ namespace System
             if ((uint)length > (uint)(array.Length - index))
                 ThrowHelper.ThrowArgumentException(ExceptionResource.Argument_InvalidOffLen);
 
-            if (length <= 1)
-                return;
-
-            ref T first = ref Unsafe.Add(ref Unsafe.As<byte, T>(ref array.GetRawSzArrayData()), index);
-            ref T last = ref Unsafe.Add(ref Unsafe.Add(ref first, length), -1);
-            do
-            {
-                T temp = first;
-                first = last;
-                last = temp;
-                first = ref Unsafe.Add(ref first, 1);
-                last = ref Unsafe.Add(ref last, -1);
-            } while (Unsafe.IsAddressLessThan(ref first, ref last));
+            MemoryMarshal.CreateSpan(ref Unsafe.Add(ref Unsafe.As<byte, T>(ref array.GetRawSzArrayData()), index), length).Reverse();
         }
 
         // Sorts the elements of an array. The sort compares the elements to each