This takes advantage of Roslyn's new ref reassignment feature to speed up the inner loop in `Array.Reverse`. Testing shows that for large arrays the reversal time is cut by approx. 20% from the baseline.

Old loop codegen:

```asm
00007ff8`8311fc2c 4c63c0          movsxd  r8,eax
00007ff8`8311fc2f 468b0c81        mov     r9d,dword ptr [rcx+r8*4]
00007ff8`8311fc33 4e8d0481        lea     r8,[rcx+r8*4]
00007ff8`8311fc37 4c63d2          movsxd  r10,edx
00007ff8`8311fc3a 468b1c91        mov     r11d,dword ptr [rcx+r10*4]
00007ff8`8311fc3e 458918          mov     dword ptr [r8],r11d
00007ff8`8311fc41 46890c91        mov     dword ptr [rcx+r10*4],r9d
00007ff8`8311fc45 ffc0            inc     eax
00007ff8`8311fc47 ffca            dec     edx
00007ff8`8311fc49 3bc2            cmp     eax,edx
00007ff8`8311fc4b 7cdf            jl      ... (00007ff8`8311fc2c)
```

New loop codegen:

```asm
00007ff8`831238e2 8b11            mov     edx,dword ptr [rcx]
00007ff8`831238e4 448b00          mov     r8d,dword ptr [rax]
00007ff8`831238e7 448901          mov     dword ptr [rcx],r8d
00007ff8`831238ea 8910            mov     dword ptr [rax],edx
00007ff8`831238ec 4883c104        add     rcx,4
00007ff8`831238f0 4883c0fc        add     rax,0FFFFFFFFFFFFFFFCh
00007ff8`831238f4 483bc8          cmp     rcx,rax
00007ff8`831238f7 72e9            jb      ... (00007ff8`831238e2)
```

The codegen samples above are for reversing an `int[]`. I saw similar performance gains for other types. Referential types like `object[]` don't get as large a benefit due to the card table checks, but there is still some speedup.

I also took the opportunity to collapse the precondition checks using the recommended patterns for (index, count) parameters.