@GrabYourPitchforks 2018-05-05
@jkotas Do you see any issue with having `Array.Reverse<T>` call into `MemoryExtensions.Reverse<T>`, as proposed in the latest commit? You've mentioned in the past that there are some weird codegen considerations, but not sure if those would apply here.