@jkotas 2018-05-06
> Do you see any issue with having Array.Reverse<T> call into MemoryExtensions.Reverse<T>

It would make `Array.Reverse<T>` slower, in particular where T is reference type or shared generic type.

> Is it possible in theory for the JIT to generate something like the below in the second case?

Yes, it is possible. It is not just theory. It is pretty likely that the JIT will generate something like this on non-Intel architectures that have less rich addressing modes.

> first and last technically point to different objects

byref that points right past the end of the an array is treated by the garbage collector as if it was pointing into the array.

From ECMA-335 spec: Managed pointers can point to ... the address where an element just past the end of an array would be stored (for pointer indexes into managed arrays.

We take advantage of this in number of places.