@jkotas 2018-05-06
@dotnet-bot test Windows_NT x64 Checked corefx_baseline