@stephentoub 2018-05-07
@dotnet-bot test Linux-musl x64 Debug Build please
@dotnet-bot test Windows_NT x64 Checked Build and Test (Jit - CoreFx) please