@GrabYourPitchforks 2018-05-07
Benchmarks!

### Old (System.Private.CoreLib as it is today)

|              Method | Size |         Mean |      Error |     StdDev |
|-------------------- |----- |-------------:|-----------:|-----------:|
|   ArrayReverse_Byte |    0 |     2.810 ns |  0.0309 ns |  0.0274 ns |
|    ArrayReverse_Int |    0 |     2.493 ns |  0.0285 ns |  0.0223 ns |
| ArrayReverse_Object |    0 |     3.265 ns |  0.0378 ns |  0.0354 ns |
|   ArrayReverse_Byte |    1 |     2.774 ns |  0.0457 ns |  0.0428 ns |
|    ArrayReverse_Int |    1 |     2.664 ns |  0.0408 ns |  0.0362 ns |
| ArrayReverse_Object |    1 |     3.446 ns |  0.0485 ns |  0.0454 ns |
|   ArrayReverse_Byte |    4 |     3.867 ns |  0.0554 ns |  0.0463 ns |
|    ArrayReverse_Int |    4 |     3.654 ns |  0.0461 ns |  0.0432 ns |
| ArrayReverse_Object |    4 |    14.045 ns |  0.1391 ns |  0.1301 ns |
|   ArrayReverse_Byte |   16 |     9.753 ns |  0.0684 ns |  0.0571 ns |
|    ArrayReverse_Int |   16 |     9.696 ns |  0.1964 ns |  0.1837 ns |
| ArrayReverse_Object |   16 |    47.152 ns |  0.3565 ns |  0.3160 ns |
|   ArrayReverse_Byte |  128 |    58.398 ns |  0.5495 ns |  0.4871 ns |
|    ArrayReverse_Int |  128 |    57.598 ns |  0.5697 ns |  0.5329 ns |
| ArrayReverse_Object |  128 |   373.269 ns |  7.2844 ns |  7.4805 ns |
|   ArrayReverse_Byte | 1024 |   441.634 ns |  4.9968 ns |  4.6740 ns |
|    ArrayReverse_Int | 1024 |   442.723 ns |  3.5002 ns |  3.2740 ns |
| ArrayReverse_Object | 1024 | 2,884.619 ns | 19.0603 ns | 15.9162 ns |

### New
(Not sure what's up with the error / stddev.)

|              Method | Size |         Mean |       Error |     StdDev |
|-------------------- |----- |-------------:|------------:|-----------:|
|   ArrayReverse_Byte |    0 |     2.184 ns |   1.8673 ns |  0.1055 ns |
|    ArrayReverse_Int |    0 |     2.072 ns |   1.6650 ns |  0.0941 ns |
| ArrayReverse_Object |    0 |     3.425 ns |   2.1635 ns |  0.1222 ns |
|   ArrayReverse_Byte |    1 |     2.296 ns |   0.7198 ns |  0.0407 ns |
|    ArrayReverse_Int |    1 |     2.309 ns |   1.4736 ns |  0.0833 ns |
| ArrayReverse_Object |    1 |     3.226 ns |   2.5964 ns |  0.1467 ns |
|   ArrayReverse_Byte |    4 |     4.137 ns |   1.4159 ns |  0.0800 ns |
|    ArrayReverse_Int |    4 |     3.365 ns |   3.5914 ns |  0.2029 ns |
| ArrayReverse_Object |    4 |    13.196 ns |   2.7496 ns |  0.1554 ns |
|   ArrayReverse_Byte |   16 |     9.395 ns |   1.1532 ns |  0.0652 ns |
|    ArrayReverse_Int |   16 |     9.242 ns |   4.6925 ns |  0.2651 ns |
| ArrayReverse_Object |   16 |    40.492 ns |   8.9226 ns |  0.5041 ns |
|   ArrayReverse_Byte |  128 |    48.476 ns |   5.5676 ns |  0.3146 ns |
|    ArrayReverse_Int |  128 |    47.164 ns |   8.2728 ns |  0.4674 ns |
| ArrayReverse_Object |  128 |   286.940 ns |  69.7246 ns |  3.9396 ns |
|   ArrayReverse_Byte | 1024 |   365.065 ns |  22.9375 ns |  1.2960 ns |
|    ArrayReverse_Int | 1024 |   348.223 ns |  43.9718 ns |  2.4845 ns |
| ArrayReverse_Object | 1024 | 2,241.796 ns | 186.5714 ns | 10.5416 ns |