@GrabYourPitchforks 2018-05-07
@jkotas It's the overload that takes only the `T[]`, but it calls the overload that takes indices. These benchmark numbers are for the first iteration of the PR before any of the `MemoryExtensions.Reverse<T>` work was performed.