@GrabYourPitchforks 2018-05-07
Unit tests are failing because of the changes to parameter validation. The PR https://github.com/dotnet/corefx/pull/29559 addresses the failing unit tests.