@jkotas 2018-05-08
@dotnet-bot test Windows_NT x64 Checked corefx_baseline