@GrabYourPitchforks 2018-05-08
I fixed most of the error / stddev issues with my benchmark runner.

### Old (System.Private.CoreLib as it is today)

|              Method | Size |         Mean |      Error |     StdDev |
|-------------------- |----- |-------------:|-----------:|-----------:|
|   ArrayReverse_Byte |    0 |     2.469 ns |  0.0243 ns |  0.0215 ns |
|    ArrayReverse_Int |    0 |     2.252 ns |  0.0871 ns |  0.1276 ns |
| ArrayReverse_Object |    0 |     2.574 ns |  0.0238 ns |  0.0172 ns |
|   ArrayReverse_Byte |    1 |     2.629 ns |  0.0272 ns |  0.0241 ns |
|    ArrayReverse_Int |    1 |     2.264 ns |  0.0122 ns |  0.0095 ns |
| ArrayReverse_Object |    1 |     3.004 ns |  0.0620 ns |  0.0550 ns |
|   ArrayReverse_Byte |    4 |     3.189 ns |  0.0819 ns |  0.0726 ns |
|    ArrayReverse_Int |    4 |     3.119 ns |  0.0401 ns |  0.0335 ns |
| ArrayReverse_Object |    4 |    11.957 ns |  0.0855 ns |  0.0668 ns |
|   ArrayReverse_Byte |   16 |     8.438 ns |  0.1104 ns |  0.0730 ns |
|    ArrayReverse_Int |   16 |     8.469 ns |  0.1493 ns |  0.1166 ns |
| ArrayReverse_Object |   16 |    42.609 ns |  0.5579 ns |  0.4034 ns |
|   ArrayReverse_Byte |  128 |    50.724 ns |  0.9750 ns |  0.8643 ns |
|    ArrayReverse_Int |  128 |    50.025 ns |  0.2537 ns |  0.2119 ns |
| ArrayReverse_Object |  128 |   331.447 ns |  5.9050 ns |  5.5235 ns |
|   ArrayReverse_Byte | 1024 |   382.151 ns |  7.6200 ns |  9.0711 ns |
|    ArrayReverse_Int | 1024 |   394.005 ns |  2.6319 ns |  2.0548 ns |
| ArrayReverse_Object | 1024 | 2,576.236 ns | 14.3080 ns | 11.9478 ns |

### New (latest iteration of the PR)


|              Method | Size |         Mean |      Error |     StdDev |
|-------------------- |----- |-------------:|-----------:|-----------:|
|   ArrayReverse_Byte |    0 |     1.933 ns |  0.0780 ns |  0.2262 ns |
|    ArrayReverse_Int |    0 |     1.994 ns |  0.0721 ns |  0.0801 ns |
| ArrayReverse_Object |    0 |     2.864 ns |  0.0885 ns |  0.0739 ns |
|   ArrayReverse_Byte |    1 |     1.710 ns |  0.0594 ns |  0.0527 ns |
|    ArrayReverse_Int |    1 |     1.901 ns |  0.0698 ns |  0.0545 ns |
| ArrayReverse_Object |    1 |     2.725 ns |  0.0854 ns |  0.1403 ns |
|   ArrayReverse_Byte |    4 |     2.919 ns |  0.0295 ns |  0.0230 ns |
|    ArrayReverse_Int |    4 |     3.316 ns |  0.0983 ns |  0.0872 ns |
| ArrayReverse_Object |    4 |    12.309 ns |  0.0822 ns |  0.0641 ns |
|   ArrayReverse_Byte |   16 |     7.793 ns |  0.0577 ns |  0.0418 ns |
|    ArrayReverse_Int |   16 |     7.392 ns |  0.1829 ns |  0.2033 ns |
| ArrayReverse_Object |   16 |    39.336 ns |  0.1710 ns |  0.1428 ns |
|   ArrayReverse_Byte |  128 |    38.962 ns |  0.2893 ns |  0.2565 ns |
|    ArrayReverse_Int |  128 |    39.529 ns |  0.5673 ns |  0.5306 ns |
| ArrayReverse_Object |  128 |   292.475 ns |  5.7809 ns |  7.5168 ns |
|   ArrayReverse_Byte | 1024 |   300.793 ns |  0.9178 ns |  0.6637 ns |
|    ArrayReverse_Int | 1024 |   301.494 ns |  1.0527 ns |  0.8791 ns |
| ArrayReverse_Object | 1024 | 2,330.756 ns | 45.4055 ns | 55.7620 ns |